+++ packages/ramps-controller/src/index.ts
@@ -8,6 +8,8 @@ export type {
   RampsControllerOptions,
   UserRegion,
   ResourceState,
+  TransakState,
+  NativeProvidersState,
 } from './RampsController';
 export {
   RampsController,
@@ -73,3 +75,47 @@ export {
 } from './RequestCache';
 export type { RequestSelectorResult } from './selectors';
 export { createRequestSelector } from './selectors';
+export type {
+  TransakServiceActions,
+  TransakServiceEvents,
+  TransakServiceMessenger,
+  TransakAccessToken,
+  TransakUserDetails,
+  TransakUserDetailsAddress,
+  TransakUserDetailsKycDetails,
+  TransakBuyQuote,
+  TransakKycRequirement,
+  TransakAdditionalRequirement,
+  TransakAdditionalRequirementsResponse,
+  TransakOttResponse,
+  TransakOrderPaymentMethod,
+  TransakDepositOrder,
+  TransakDepositNetwork,
+  TransakDepositCryptoCurrency,
+  TransakDepositPaymentMethod,
+  TransakDepositRegion,
+  TransakOrder,
+  TransakQuoteTranslation,
+  TransakTranslationRequest,
+  TransakUserLimits,
+  TransakIdProofStatus,
+  PatchUserRequestBody as TransakPatchUserRequestBody,
+} from './TransakService';
+export {
+  TransakApiError,
+  TransakService,
+  TransakEnvironment,
+  TransakOrderIdTransformer,
+} from './TransakService';
+export type {
+  TransakServiceMethodActions,
+  TransakServiceSendUserOtpAction,
+  TransakServiceVerifyUserOtpAction,
+  TransakServiceGetUserDetailsAction,
+  TransakServiceGetBuyQuoteAction,
+  TransakServiceGetKycRequirementAction,
+  TransakServiceCreateOrderAction,
+  TransakServiceGetOrderAction,
+  TransakServiceRequestOttAction,
+  TransakServiceGeneratePaymentWidgetUrlAction,
+} from './TransakService-method-action-types';